Cleanup of the dead end solution
Random Access Memory Buffer works for libstdc++ because it writes and moves the read buffer's end at the same time.
For libc++ implementation (based on iterators) write is performed and only pptr is adjusted. That messes up situation when write is performed in the middle of the stream, as we loose info where actual data ends (provided we overallocate). I believe I don't quite get iostream's logic. It seems I missed something.
For now we'll keep the simples solution with file buffer.